Version 0.5.4, database tests
