+++ auth/src/test/java/com/firebase/ui/auth/util/HeadlessAPIWrapperImplTest.java
@@ -14,10 +14,6 @@
 
 package com.firebase.ui.auth.util;
 
-import static junit.framework.Assert.assertFalse;
-import static junit.framework.Assert.assertTrue;
-import static org.mockito.Mockito.when;
-
 import com.firebase.ui.auth.BuildConfig;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
@@ -32,8 +28,12 @@ import org.robolectric.RobolectricGradleTestRunner;
 import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Config;
 
+import static junit.framework.Assert.assertFalse;
+import static junit.framework.Assert.assertTrue;
+import static org.mockito.Mockito.when;
+
 @RunWith(RobolectricGradleTestRunner.class)
-@Config(constants = BuildConfig.class)
+@Config(constants = BuildConfig.class, sdk=23)
 public class HeadlessAPIWrapperImplTest {
 
     private FirebaseAuthWrapperImpl mHeadlessAPIWrapperImpl;
