+++ common/constants.gradle
@@ -1,9 +1,9 @@
-project.ext.firebase_version = '9.4.0'
+project.ext.firebase_version = '9.6.0'
 project.ext.support_library_version = '23.4.0'
 
 project.ext.submodules = ['database', 'auth']
 project.ext.group = "com.firebaseui"
-project.ext.version = '0.5.3'
+project.ext.version = '0.5.4'
 project.ext.pomdesc = 'Firebase UI Android'
 project.ext.buildtools = '23.0.3'
 project.ext.compileSdk = 23
