+++ database/src/androidTest/java/com/firebase/ui/database/FirebaseArrayTest.java
@@ -14,7 +14,9 @@
 
 package com.firebase.ui.database;
 
-import android.test.AndroidTestCase;
+import android.support.test.runner.AndroidJUnit4;
+import android.test.InstrumentationTestCase;
+import android.test.suitebuilder.annotation.SmallTest;
 
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.database.DatabaseError;
@@ -27,72 +29,48 @@ import junit.framework.AssertionFailedError;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 
-public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
-    public static class Bean {
-        int number;
-        String text;
-        boolean bool;
+@RunWith(AndroidJUnit4.class)
+@SmallTest
+public class FirebaseArrayTest extends InstrumentationTestCase {
 
-        public Bean() {
-            // necessary for Jackson
-        }
-
-        public Bean(int number, String text, boolean bool) {
-            this.number = number;
-            this.text = text;
-            this.bool = bool;
-        }
-        public Bean(int index) {
-            this(index, "Text "+index, index % 2 == 0);
-        }
-
-        public int getNumber() {
-            return number;
-        }
-
-        public String getText() {
-            return text;
-        }
-
-        public boolean isBool() {
-            return bool;
-        }
-    }
+    private static final int TIMEOUT = 5000;
 
     private DatabaseReference mRef;
     private FirebaseArray mArray;
 
     @Before
     public void setUp() throws Exception {
-        FirebaseApp app = ApplicationTest.getAppInstance(getContext());
-        mRef = FirebaseDatabase.getInstance(app).getReference()
-                .child("firebasearray").child("objects");
+        FirebaseApp app = ApplicationTest.getAppInstance(getInstrumentation().getContext());
+        mRef = FirebaseDatabase.getInstance(app).getReference().child("firebasearray");
         mArray = new FirebaseArray(mRef);
         mRef.removeValue();
         runAndWaitUntil(mArray, mRef, new Runnable() {
-                    @Override
-                    public void run() {
-                        for (int i = 1; i <= 3; i++) {
-                            mRef.push().setValue(new Bean(i, "Text " + i, i % 2 == 0 ? true : false), i);
-                        }
-                    }
-                }, new Callable<Boolean>() {
-                    @Override
-                    public Boolean call() throws Exception {
-                        return mArray.getCount() == 3;
-                    }
+            public void run() {
+                for (int i = 1; i <= 3; i++) {
+                    mRef.push().setValue(i, i);
                 }
-        );
+            }
+        }, new Callable<Boolean>() {
+            public Boolean call() throws Exception {
+                return mArray.getCount() == 3;
+            }
+        });
     }
 
     @After
     public void tearDown() throws Exception {
-        mRef.removeValue();
-        mArray.cleanup();
+        if (mRef != null) {
+            mRef.removeValue();
+        }
+
+        if (mArray != null) {
+            mArray.cleanup();
+        }
     }
 
     @Test
@@ -105,7 +83,7 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
         assertEquals(3, mArray.getCount());
         runAndWaitUntil(mArray, mRef, new Runnable() {
             public void run() {
-                mRef.push().setValue(new Bean(4));
+                mRef.push().setValue(4);
             }
         }, new Callable<Boolean>() {
             @Override
@@ -118,12 +96,12 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
     public void testPushAppends() throws Exception {
         runAndWaitUntil(mArray, mRef, new Runnable() {
             public void run() {
-                mRef.push().setValue(new Bean(4), 4);
+                mRef.push().setValue(4, 4);
             }
         }, new Callable<Boolean>() {
             @Override
             public Boolean call() throws Exception {
-                return mArray.getItem(3).getValue(Bean.class).getNumber() == 4;
+                return mArray.getItem(3).getValue(Integer.class).equals(4);
             }
         });
     }
@@ -132,11 +110,11 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
     public void testAddValueWithPriority() throws Exception {
         runAndWaitUntil(mArray, mRef, new Runnable() {
             public void run() {
-                mRef.push().setValue(new Bean(4), 0.5);
+                mRef.push().setValue(4, 0.5);
             }
         }, new Callable<Boolean>() {
             public Boolean call() throws Exception {
-                return mArray.getItem(3).getValue(Bean.class).getNumber() == 3 && mArray.getItem(0).getValue(Bean.class).getNumber() == 4;
+                return mArray.getItem(3).getValue(Integer.class).equals(3) && mArray.getItem(0).getValue(Integer.class).equals(4);
             }
         });
     }
@@ -149,8 +127,7 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
             }
         }, new Callable<Boolean>() {
             public Boolean call() throws Exception {
-                return getBean(mArray, 0).getNumber() == 3 && getBean(mArray, 1).getNumber() == 1 && getBean(mArray, 2).getNumber() == 2;
-                //return isValuesEqual(mArray, new int[]{3, 1, 2});
+                return isValuesEqual(mArray, new int[]{3, 1, 2});
             }
         });
     }
@@ -165,8 +142,15 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
         return true;
     }
 
-    private Bean getBean(FirebaseArray array, int index) {
-        return array.getItem(index).getValue(Bean.class);
+    private Integer getIntValue(FirebaseArray array, int index) {
+        return array.getItem(index).getValue(Integer.class);
+    }
+
+    private static void print(FirebaseArray array) {
+        for (int i=0; i < array.getCount(); i++) {
+            System.out.println(i+": key="+array.getItem(i).getKey()+" value="+array.getItem(i).getValue());
+
+        }
     }
 
     public static void runAndWaitUntil(final FirebaseArray array, Query ref, Runnable task, Callable<Boolean> done) throws InterruptedException {
@@ -184,7 +168,7 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
         task.run();
         boolean isDone = false;
         long startedAt = System.currentTimeMillis();
-        while (!isDone && System.currentTimeMillis() - startedAt < 5000) {
+        while (!isDone && System.currentTimeMillis() - startedAt < TIMEOUT) {
             semaphore.tryAcquire(1, TimeUnit.SECONDS);
             try {
                 isDone = done.call();
@@ -197,4 +181,5 @@ public class FirebaseArrayOfObjectsTest extends AndroidTestCase {
             throw new AssertionFailedError();
         }
         array.setOnChangedListener(null);
-    }}
+    }
+}
